chore: move guards and strategies in folders
