+++ src/auth/auth.module.ts
@@ -4,11 +4,11 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { UserModule } from 'src/user/user.module';
-import { LocalStrategy } from './local.strategy';
-import { JwtStrategy } from './jwt.strategy';
+import { LocalStrategy } from './strategy/local.strategy';
+import { JwtStrategy } from './strategy/jwt.strategy';
 import { APP_GUARD } from '@nestjs/core';
-import { JwtAuthGuard } from './jwt-auth.guard';
-import { RefreshStrategy } from './refresh.strategy';
+import { JwtAuthGuard } from './guards/jwt-auth.guard';
+import { RefreshStrategy } from './strategy/refresh.strategy';
 
 @Module({
   imports: [
